Timezone conversion is not present in UI for 2.2
Magento 2.1: in table cron_schedule stored task timestamps in client time zone. For example New York time zone.
Magento 2.2: in table cron_schedule stored task timestamps in UTC.
As the result on Magento 2.2, on the page "tasks list" all dates displayed in UTC instead of client timezone.

Timeline rendered in client time zone.

This is the fix. However, it's most likely not compatible with 2.1 and 2.0 (I did not test it)